# openapi.yaml
openapi: 3.1.0
info:
  title: ACAS Pro API
  description: |
    ACAS Pro 企业级智能营销系统 API
    
    ## 功能模块
    - 用户认证与授权
    - 销售预测引擎
    - 库存优化
    - 市场情报监测
    - 内容创作引擎
    - 多平台账号管理
    
    ## 认证方式
    使用 JWT Bearer Token 认证，在请求头中添加：
    ```
    Authorization: Bearer <your_access_token>
    ```
    
    ## 错误码
    - 400: 请求参数错误
    - 401: 未授权/Token过期
    - 403: 权限不足
    - 404: 资源不存在
    - 429: 请求频率超限
    - 500: 服务器内部错误
    
  version: 4.0.0
  contact:
    name: ACAS Pro Support
    email: support@acas-pro.example.com
  license:
    name: Enterprise License
    url: https://acas-pro.example.com/license
servers:
  - url: http://localhost:8000
    description: 本地开发服务器
  - url: https://api.acas-pro.example.com
    description: 生产环境
...
paths:
...
  /api/v1/auth/login:
    post:
      tags: [auth]
      summary: 用户登录
      description: 使用账号密码登录，获取访问令牌
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [account, password]
              properties:
                account:
                  type: string
                  description: 用户账号
                password:
                  type: string
                  description: 密码
            example:
              account: testuser
              password: SecurePass123
      responses:
        '200':
          description: 登录成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token:
                    type: string
                    description: 访问令牌
                  refresh_token:
                    type: string
                    description: 刷新令牌
                  token_type:
                    type: string
                    default: Bearer
                  expires_in:
                    type: integer
                    description: 令牌有效期（秒）
                  user:
                    $ref: '#/components/schemas/UserProfile'
        '401':
          description: 账号或密码错误
        '429':
          description: 登录尝试次数过多，账户已锁定
...
  /api/v1/users/me:
    get:
      tags: [users]
      summary: 获取当前用户信息
      security:
        - BearerAuth: []
      responses:
        '200':
          description: 用户信息
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'

    put:
      tags: [users]
      summary: 更新用户信息
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                nickname:
                  type: string
                email:
                  type: string
                  format: email
                phone:
                  type: string
                language:
                  type: string
                  enum: [zh-CN, en-US, zh-TW]
                timezone:
                  type: string
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT 访问令牌
  schemas:
    UserProfile:
      type: object
      properties:
        id:
          type: string
        account:
          type: string
        nickname:
          type: string
        email:
          type: string
        phone:
          type: string
        role:
          type: string
          enum: [admin, user, viewer]
        status:
          type: string
          enum: [active, inactive, locked]
        region:
          type: string
        language:
          type: string
        timezone:
          type: string
        created_at:
          type: string
          format: date-time
        last_login:
          type: string
          format: date-time
        wallet_balance:
          type: number
        model_preference:
          type: string
...